+++ packages/imperative/src/imperative/src/Imperative.ts
@@ -13,7 +13,6 @@
  * Main class of the Imperative framework, returned when you
  * require("@zowe/imperative") e.g. const imperative =  require("@zowe/imperative");
  */
-import { PerfTiming } from "@zowe/perf-timing";
 import { Logger, LoggerConfigBuilder } from "../../logger";
 import { IImperativeConfig } from "./doc/IImperativeConfig";
 import { Arguments } from "yargs";
@@ -59,22 +58,6 @@ import { ISettingsFile } from "../../settings/src/doc/ISettingsFile";
 import { CompleteAuthGroupBuilder } from "./auth/builders/CompleteAuthGroupBuilder";
 import { ICommandProfileAuthConfig } from "../../cmd/src/doc/profiles/definition/ICommandProfileAuthConfig";
 import { ImperativeExpect } from "../../expect";
-
-// Bootstrap the performance tools
-if (PerfTiming.isEnabled) {
-    // These are expensive operations so imperative should
-    // only do it when performance is enabled.
-    const Module = require("module");
-
-    // Store the reference to the original require.
-    const originalRequire = Module.prototype.require;
-
-    // Timerify a wrapper named function so we can be sure that not just
-    // any anonymous function gets checked.
-    Module.prototype.require = PerfTiming.api.watch(function NodeModuleLoader(...args: any[]) {
-        return originalRequire.apply(this, args);
-    });
-}
 
 export class Imperative {
 
@@ -126,13 +109,6 @@ export class Imperative {
         // eslint-disable-next-line no-async-promise-executor
         return new Promise<void>(async (initializationComplete: () => void, initializationFailed: ImperativeReject): Promise<void> => {
             try {
-
-                const timingApi = PerfTiming.api;
-
-                if (PerfTiming.isEnabled) {
-                    // Marks point START
-                    timingApi.mark("START_IMP_INIT");
-                }
 
                 /**
                  * Config Logger Manager to enable log messages in memory prior to logger init.
@@ -254,12 +230,6 @@ export class Imperative {
                  */
                 initializationComplete();
 
-                if (PerfTiming.isEnabled) {
-                    // Marks point END
-                    timingApi.mark("END_IMP_INIT");
-                    timingApi.measure("Imperative.init()", "START_IMP_INIT", "END_IMP_INIT");
-                }
-
             } catch (error) {
                 const imperativeLogger = Logger.getImperativeLogger();
                 imperativeLogger.fatal(require("util").inspect(error));
@@ -303,21 +273,7 @@ export class Imperative {
      * @returns {Imperative} this, for chaining syntax
      */
     public static parse(): Imperative {
-
-        const timingApi = PerfTiming.api;
-
-        if (PerfTiming.isEnabled) {
-            // Marks point START
-            timingApi.mark("START_IMP_PARSE");
-        }
-
         Imperative.yargs.parse();
-
-        if (PerfTiming.isEnabled) {
-            // Marks point END
-            timingApi.mark("END_IMP_PARSE");
-            timingApi.measure("Imperative.init()", "START_IMP_PARSE", "END_IMP_PARSE");
-        }
         return this;
     }